Make AliasAwareOutputPartitioning trait
